feat(bpm): add input support for typing BPM in useBPM hook and ControlBPM component
